添加记录购买、持仓、收益的历史记录

+++ src/model_db/tbl_operation_record.py
@@ -1,6 +1,6 @@
 from datetime import datetime
 from sqlalchemy.ext.declarative import declarative_base
-from sqlalchemy import Column, DateTime, DECIMAL, VARCHAR, CHAR, INT
+from sqlalchemy import Column, DateTime, DECIMAL, VARCHAR, CHAR, BIGINT
 
 Base = declarative_base()
 
@@ -9,7 +9,7 @@ class OperationRecordTable(Base):
 
     __tablename__ = 'tbl_operation_record'
 
-    id = Column(INT, primary_key=True, autoincrement=True)
+    id = Column(BIGINT, primary_key=True, autoincrement=True)
     name = Column(VARCHAR(255), nullable=False)
     code = Column(CHAR(6), nullable=False)
     operate_type = Column(VARCHAR(32), nullable=False)
